Restructured directories to better fit discord like layout
