+++ frontend/src/components/Navbar/Navbar.css
@@ -1,0 +1,18 @@
+.navbar-container {
+    height: 100vh;
+    width: 72px;
+    background-color: #202225;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 12px 0px;
+    gap: 10px;
+}
+
+.navbar-server-icon {
+    height: 48px;
+    width: 48px;
+    border-radius: 50%;
+    object-fit: cover;
+    cursor: pointer;
+}
