--- frontend/src/components/ServerSideBar/ServerSideBar.css
@@ -1,5 +1,0 @@
-.server-side-bar-container {
-    height: 100vh;
-    width: 72px;
-    background-color: #202225;
-}
