--- frontend/src/components/ServerSideBar/index.js
@@ -1,18 +1,0 @@
-import './ServerSidebar.css';
-import ServerSidebarCard from '../ServerSidebarCard';
-import { useSelector } from 'react-redux';
-
-const ServerSidebar = () => {
-    const servers = useSelector(state => state.serversState);
-    const serversArr = Object.values(servers);
-
-    return (
-        <div className='server-side-bar-container'>
-            {serversArr.map(server => (
-                <ServerSidebarCard server={server} key={server.id} />
-            ))}
-        </div>
-    );
-};
-
-export default ServerSidebar;
