--- frontend/src/components/ServerSideBarCard/index.js
@@ -1,12 +1,0 @@
-import './ServerSidebarCard.css';
-
-const ServerSidebarCard = ({server}) => {
-    console.log(server.iconURL)
-    return (
-        <div>
-            <img src={server.iconURL} alt='icon' />
-        </div>
-    );
-};
-
-export default ServerSidebarCard;
