+++ frontend/src/components/Sidebar/index.js
@@ -1,0 +1,12 @@
+import './Sidebar.css';
+
+const Sidebar = () => {
+
+    return (
+        <div>
+            Sidebar Component
+        </div>
+    );
+};
+
+export default Sidebar;
