+++ frontend/src/components/UserApplication/index.js
@@ -1,0 +1,13 @@
+import './UserApplication.css';
+import ServerSidebar from '../Navbar';
+import { useState } from 'react';
+
+const UserApplication = () => {
+    const [currentServer, setCurrentServer] = useState('');
+
+    return (
+        <ServerSidebar setCurrentServer={setCurrentServer} />
+    );
+};
+
+export default UserApplication;
